# django>=1.8 required, speed up bulk_update, new feature