--- a/requirements-test.txt
+++ b/requirements-test.txt
@@ -1,5 +1,4 @@
 dj_database_url==0.4.2
-django-uuidfield==0.5.0
 jsonfield==2.0.1
 pillow==4.1.1
 six==1.10.0