--- a/runtest.py
+++ b/runtest.py
@@ -9,8 +9,7 @@ if __name__ == "__main__":
     from django.conf import settings
     from django.test.utils import get_runner
 
-    if django.VERSION >= (1, 7):
-       django.setup()
+    django.setup()
 
     TestRunner = get_runner(settings)
     test_runner = TestRunner()