--- a/tests/requirements/django-1.5.txt
+++ b/tests/requirements/django-1.5.txt
@@ -1,3 +1,0 @@
--r requirements_base.txt
-django>=1.5,<1.6
-django-uuidfield==0.5.0