--- a/tox.ini
+++ b/tox.ini
@@ -1,17 +1,15 @@
 [tox]
 envlist =
-    py{27,34}-django{15,16,17,18,19}
+    py{27,35}-django{18,19,110}
 
 [testenv]
 setenv =
     PYTHONDONTWRITEBYTECODE=1
 install_command = pip install {opts} {packages}
 deps =
-    django15: Django>=1.5,<1.6
-    django16: Django>=1.6,<1.7
-    django17: Django>=1.7,<1.8
     django18: Django>=1.8,<1.9
     django19: Django>=1.9,<1.10
+    django110: Django>=1.10,<1.11
     -rrequirements-test.txt
 commands = ./runtest.py {posargs}