… bulk_update/__init__.py
… django_bulk_update/__init__.py
… django_bulk_update/manager.py